Move accumulator code to its own package.
